Добавили вывод сообщений после submit
